fix: Support serialize Gradient.colorSpace
