+++ Alchemy/Assets/Alchemy/Tests/Runtime/GradientSerializationTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c218e05a8a5b40b59cc115c252081e14
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
